+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 **Unreleased**
 --------------
 
+- **New**: Add `ComposeNonRestartableComposable` to suggest `@NonRestartableComposable` for composables that only forward to one other composable.
+
 1.5.5
 -----